Refactor project structure and update imports
- Updated import statements across multiple files to reflect new module structure.
- Removed unnecessary `config` and `db_utils` subdirectory references in imports.
- Changed development dependencies in requirements.txt to be uncommented for easier setup.
- Deleted outdated setup.py script and replaced it with a new configuration file (config.py) for better management of environment variables and database configurations.
- Introduced a new database utility module (db_utils.py) to handle database connections and operations.
- Removed deprecated user utilities (user_utils.py) and replaced with a more streamlined version.
- Cleaned up utility functions in utils.py and removed unused functions.
- Ensured all migration-related modules are properly organized and functional.
