+++ migrators/teacher_migrator.py
@@ -7,8 +7,8 @@ import logging
 from typing import Dict, Any, List, Optional
 from datetime import datetime
 
-from ..db_utils.db_utils import db_manager
-from ..user_utils.user_utils import user_manager
+from ..db_utils import db_manager
+from ..user_utils import user_manager
 from .school_migrator import school_migrator
 
 logger = logging.getLogger(__name__)
