+++ requirements.txt
@@ -24,10 +24,10 @@ tabulate>=0.9.0
 
 # Development dependencies (optional)
 # Uncomment for development
-# pytest>=7.4.0
-# pytest-asyncio>=0.21.0
-# pytest-cov>=4.1.0
-# black>=23.0.0
-# isort>=5.12.0
-# flake8>=6.0.0
-# mypy>=1.5.0
+pytest>=7.4.0
+pytest-asyncio>=0.21.0
+pytest-cov>=4.1.0
+black>=23.0.0
+isort>=5.12.0
+flake8>=6.0.0
+mypy>=1.5.0
