+++ user_utils.py
@@ -9,8 +9,8 @@ from typing import Dict, Any, Optional, List
 from dataclasses import dataclass
 from enum import Enum
 
-from ..db_utils.db_utils import db_manager
-from ..config.config import DEFAULT_VALUES
+from ..db_utils import db_manager
+from ..config import DEFAULT_VALUES
 
 logger = logging.getLogger(__name__)
